Completed the coding needed for the urls and views files within the aquariums, food, and fish apps + added the food, plants, and water parameters' paths to the main app's urls file + made a field change within the water parameters model from pH being stored as an integer to being stored as a decimal and migrated the changes to the database.

+++ drf_jwt_capstone_backend/water_parameters/models.py
@@ -10,7 +10,7 @@ class Water_Parameter(models.Model):
     aquarium = models.ForeignKey(Aquarium, on_delete=models.CASCADE)
     general_hardness_ppm = models.IntegerField()
     carbonate_hardness_ppm = models.IntegerField()
-    power_of_hydrogen = models.IntegerField()
+    power_of_hydrogen = models.DecimalField(max_digits = 2,decimal_places=1)
     nitrites_ppm = models.IntegerField()
     nitrates_ppm = models.IntegerField()
     temperature_fahrenheit = models.IntegerField()
